added search users function and update password function for users

+++ app/Http/Controllers/backend/UserController.php
@@ -16,9 +16,13 @@ class UserController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
         $users = User::all();
+
+        if($request->has('search')) {
+            $users = User::where('username', 'like', "%{$request->search}%")->orWhere('email', 'like', "%{$request->search}%")->get();
+        }
         return view('users.index',compact('users'));
     }
 
@@ -89,6 +93,9 @@ class UserController extends Controller
      */
     public function destroy(User $user )
     {
+        if(auth()->user()->id == $user->id) {
+            return redirect()->route('user.index')->with('message', 'you can\'t perform this operation on your current account' );
+        }
         $user->delete();
 
         return redirect()->route('user.index')->with('message', 'user deleted successfully');
